+++ CHANGELOG.md
@@ -8,13 +8,14 @@ All notable changes to the Pony compiler and standard library will be documented
 
 - Avoid fairly easy to trigger overflow in Windows Time.nanos code ([PR #4227](https://github.com/ponylang/ponyc/pull/4227))
 - Fix incorrect interaction between String/Array reserve and Pointer realloc ([PR #4223](https://github.com/ponylang/ponyc/pull/4223))
+- Fix broken documentation generation on Windows ([PR #4226](https://github.com/ponylang/ponyc/pull/4226))
 
 ### Added
-
 
 ### Changed
 
 - Sort package types in documentation ([PR #4228](https://github.com/ponylang/ponyc/pull/4228))
+- Adapt documentation generation to use the mkdocs material theme ([PR #4226](https://github.com/ponylang/ponyc/pull/4226)
 
 ## [0.51.4] - 2022-10-29